Switch float in double

+++ SE2021_Gruppo5/test/it/unisa/diem/se/group5/calculator/complex/ComplexNumberTest.java
@@ -47,12 +47,12 @@ public class ComplexNumberTest {
         ComplexNumber instance1 = new ComplexNumber();
         ComplexNumber instance2 = new ComplexNumber(3.1234599f,10);
         ComplexNumber instance3 = new ComplexNumber(-3);
-        float expResult1 = 0;
-        float expResult2 = 3.12346f;
-        float expResult3 = -3;
-        float result1 = instance1.getReal();
-        float result2 = instance2.getReal();
-        float result3 = instance3.getReal();
+        double expResult1 = 0;
+        double expResult2 = 3.12346f;
+        double expResult3 = -3;
+        double result1 = instance1.getReal();
+        double result2 = instance2.getReal();
+        double result3 = instance3.getReal();
         assertEquals(expResult1, result1, 0.0);
         assertEquals(expResult2, result2, 0.0);
         assertEquals(expResult3, result3, 0.0);
@@ -65,7 +65,7 @@ public class ComplexNumberTest {
     @Test
     public void testSetReal() {
         System.out.println("setReal");
-        float real = 0.0F;
+        double real = 0.0F;
         ComplexNumber instance = new ComplexNumber();
         instance.setReal(real);
     }
@@ -79,12 +79,12 @@ public class ComplexNumberTest {
         ComplexNumber instance1 = new ComplexNumber();
         ComplexNumber instance2 = new ComplexNumber(3.777823f,-10.202f);      
         ComplexNumber instance3 = new ComplexNumber(3);
-        float expResult1 = 0;
-        float expResult2 = -10.202f;
-        float expResult3 = 0;
-        float result1 = instance1.getImaginary();
-        float result2 = instance2.getImaginary();
-        float result3 = instance3.getImaginary();
+        double expResult1 = 0;
+        double expResult2 = -10.202f;
+        double expResult3 = 0;
+        double result1 = instance1.getImaginary();
+        double result2 = instance2.getImaginary();
+        double result3 = instance3.getImaginary();
         assertEquals(expResult1, result1, 0);
         assertEquals(expResult2, result2, 0);
         assertEquals(expResult3, result3, 0);
@@ -96,7 +96,7 @@ public class ComplexNumberTest {
     @Test
     public void testSetImaginary() {
         System.out.println("setImaginary");
-        float imaginary = 0.0F;
+        double imaginary = 0.0F;
         ComplexNumber instance = new ComplexNumber();
         instance.setImaginary(imaginary);
     }
@@ -177,7 +177,7 @@ public class ComplexNumberTest {
      */
     @Test (expected =  NumberFormatException.class)
     public void testNumberFormatrException() {
-        ComplexNumber op1 = new ComplexNumber((float) Math.pow(2, 130),10);
+        ComplexNumber op1 = new ComplexNumber(Math.pow(2, 130),10);
     }
     
 }
